requirements.txt
argon2-cffi==20.1.0
cython==0.29.21
django-cors-headers==3.6.0
django-cors-middleware==1.5.0
django-token==0.1.3
djangorestframework==3.12.2
docopt==0.6.2
imageio==2.8.0
jaraco-functools==3.1.0
more-itertools==8.6.0
numpy==1.18.4
pipreqs==0.4.10
pyastar @ git+git://github.com/hjweide/pyastar.git@master#egg=pyastar
yarg==0.1.9